+++ packages/builder/builder-webpack-provider/src/config/validate/source.ts
@@ -1,0 +1,5 @@
+import { sharedSourceConfigSchema, z } from '@modern-js/builder-shared';
+import type { SourceConfig } from '../../types';
+
+export const sourceConfigSchema: z.ZodType<SourceConfig> =
+  sharedSourceConfigSchema;