+++ packages/builder/builder-webpack-provider/src/types/config/tools.ts
@@ -3,6 +3,7 @@ import type {
   ArrayOrNot,
   ChainedConfig,
   SharedToolsConfig,
+  FileFilterUtil,
 } from '@modern-js/builder-shared';
 import type {
   ModifyWebpackChainUtils,
@@ -42,10 +43,7 @@ export type ToolsBabelConfig = ChainedConfig<
 
 export type ToolsTSLoaderConfig = ChainedConfig<
   TSLoaderOptions,
-  {
-    addIncludes: (items: string | RegExp | (string | RegExp)[]) => void;
-    addExcludes: (items: string | RegExp | (string | RegExp)[]) => void;
-  }
+  { addIncludes: FileFilterUtil; addExcludes: FileFilterUtil }
 >;
 
 export type ToolsStyledComponentConfig = ChainedConfig<IStyledComponentOptions>;
@@ -62,21 +60,19 @@ export type ToolsAutoprefixerConfig = ChainedConfig<AutoprefixerOptions>;
 
 export type ToolsPostCSSLoaderConfig = ChainedConfig<
   PostCSSLoaderOptions,
-  {
-    addPlugins: (plugins: PostCSSPlugin | PostCSSPlugin[]) => void;
-  }
+  { addPlugins: (plugins: PostCSSPlugin | PostCSSPlugin[]) => void }
 >;
 
 export type ToolsPugConfig = true | ChainedConfig<PugOptions>;
 
 export type ToolsLessConfig = ChainedConfig<
   LessLoaderOptions,
-  { addExcludes: (excludes: RegExp | RegExp[]) => void }
+  { addExcludes: FileFilterUtil }
 >;
 
 export type ToolsSassConfig = ChainedConfig<
   SassLoaderOptions,
-  { addExcludes: (excludes: RegExp | RegExp[]) => void }
+  { addExcludes: FileFilterUtil }
 >;
 
 export type ToolsTSCheckerConfig = ChainedConfig<ForkTSCheckerOptions>;