+++ packages/cli/doc-core/src/theme-default/components/Search/index.tsx
@@ -4,6 +4,7 @@ import LoadingSvg from './assets/loading.svg';
 import type { MatchResultItem, PageSearcher } from './logic/search';
 import styles from './index.module.scss';
 import { SuggestionContent } from './Suggestion';
+import { usePageData } from '@/runtime';
 
 const KEY_CODE = {
   ARROW_UP: 'ArrowUp',
@@ -29,6 +30,7 @@ export function Search(props: SearchOptions) {
   const initPageSearcherPromiseRef = useRef<Promise<void>>();
   const [disableInput, setDisableInput] = useState(true);
   const searchInputRef = useRef<HTMLInputElement | null>(null);
+  const { siteData } = usePageData();
   // initializing or searching
   const showLoading = !initialized || searching;
   // 1. page searcher has been initialized and finish searching
@@ -43,6 +45,7 @@ export function Search(props: SearchOptions) {
         defaultLang,
         langs,
         langRoutePrefix,
+        pages: siteData.pages,
       });
       await psRef.current.init();
       setInitialized(true);
@@ -165,13 +168,13 @@ export function Search(props: SearchOptions) {
       </div>
       {focused && query && (
         <ul
-          pos="fixed sm:absolute top-12 sm:top-8 left-0"
+          pos="fixed sm:absolute top-12 sm:top-8 right--10"
           z="60"
           border-1=""
           p="2"
           list="none"
           bg="bg-default"
-          className="w-100% sm:min-w-500px sm:max-w-700px"
+          className="w-100% sm:min-w-600px sm:max-w-800px"
         >
           {/* Show the suggestions */}
           {suggestions.map((item, index) => (