+++ packages/solutions/module-tools/tests/fixtures/sourcemap/inline.ts
@@ -3,6 +3,6 @@ import { defineConfig } from '@modern-js/self/defineConfig';
 export default defineConfig({
   buildConfig: {
     sourceMap: 'inline',
-    outdir: './dist/inline',
+    outDir: './dist/inline',
   },
 });