+++ packages/server/plugin-koa/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @modern-js/plugin-koa
 
+## 2.0.0-beta.7
+
+### Major Changes
+
+- dda38c9c3e: chore: v2
+
+### Patch Changes
+
+- Updated dependencies [edd1cfb1af]
+- Updated dependencies [6bda14ed71]
+- Updated dependencies [dda38c9c3e]
+- Updated dependencies [8b8e1bb571]
+- Updated dependencies [bbe4c4ab64]
+  - @modern-js/utils@2.0.0-beta.7
+  - @modern-js/types@2.0.0-beta.7
+  - @modern-js/bff-core@2.0.0-beta.7
+  - @modern-js/bff-runtime@2.0.0-beta.7
+
 ## 2.0.0-beta.6
 
 ### Major Changes